refactor(web): extract HTTP transport layer from real-api
Separate the generic fetch logic (URL construction, session management,
error handling) into a dedicated transport module. This makes the
transport mechanism independently configurable and testable, while
real-api.ts focuses purely on endpoint definitions.
